> api.dart added for api.v1 handler and to echo handler > purge.dart added and supporting isActive, isRunning > server.dart arranged
